requirements.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.13
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-user-tracking==0.4.11
django-user-visit==1.1
django-utils-six==2.0
idna==3.4
IP2Location==8.9.0
mysql-connector-python==8.0.32
Pillow==8.3.1
protobuf==3.20.3
PyMySQL==1.0.2
pytz==2022.1
requests==2.28.2
sqlparse==0.4.2
typing_extensions==4.1.1
ua-parser==0.16.1
urllib3==1.26.14
user-agents==2.2.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
Django==3.2.7
django-crispy-forms==1.12.0
idna==3.2
lxml==4.6.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
python-pptx==0.6.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
XlsxWriter==3.0.1
